+++ src/common/btree/btree_root_pointer.h
@@ -1,3 +1,18 @@
+/**
+ * (C) 2010-2011 Alibaba Group Holding Limited.
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * version 2 as published by the Free Software Foundation.
+ *
+ * Version: $Id$
+ *
+ * ./btree_root_pointer.h for ...
+ *
+ * Authors:
+ *   duolong <duolong@taobao.com>
+ *
+ */
 #ifndef OCEANBASE_COMMON_BTREE_BTREE_ROOT_POINTER_H_
 #define OCEANBASE_COMMON_BTREE_BTREE_ROOT_POINTER_H_
 
@@ -74,7 +89,7 @@ class BtreeRootPointer {
   /**
    * copy node数量
    */
-  int32_t get_copy_node_count();
+  size_t get_copy_node_count();
 
   /**
    * reset copy list
@@ -183,4 +198,5 @@ class BtreeRootPointerList {
 } // end namespace sb
 
 #endif
+
